Add files via upload
Note: All images were pulled from google image search and could be protected images. These images are just for prototyping and proof of concept for the image filters in index.js. Any image claimed to belong to someone will be removed immediately upon request. 

Initial push to publish. There are some issues when trying to change between images and where the images are shown on the screen. Still in prototyping phase, as a result, UI is as basic as it can get. 

Planned future release features:
- more complete documentation and comments
- create an actual Image Filter Class instead of just a collection of functions
- will include a Filter Stack Object so computational composition can be achieved
- will incorporate more filters that will take advantage of the 3Dimensional array data structure
- will include contour line detection
- black and white filter
- alpha channel clipping masks
- comic book filter (flat color regions with darkened edges to emphasize shapes)
- many many more...